WIP Move fetch.js from lib/background to lib/utils
See https://github.com/OctoLinker/browser-extension/pull/306#discussion_r104538211
