+++ lib/click-handler.js
@@ -1,7 +1,7 @@
 import $ from 'jquery';
 import { showTooltip } from './gh-interface.js';
 import * as storage from './options/storage.js';
-import fetch from './background/fetch.js';
+import fetch from './utils/fetch.js';
 
 const SORRY = 'I\'m sorry, unable to resolve this link  😱';
 const PROCESS = 'Processing  ⏳';
